refactor: useRestaurants 모듈 이동 및 경로 수정
